# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IZ-GET",
    "description": "코드 분석 및 소크라틱 문답과 채점, 교안 분석 및 안내를 위한 AI 서비스",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v0/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "코드 분석 요청 (P02)",
        "description": "분석 요청 접수하고 즉시 202 반환. 실제 분석은 백그라운드\n\nx_trace_id는 아직 사용 X. 헤더 자리 게약 고정하고 Swagger 노출 위해 지금 받아둠\n- 로깅에 붙이는 것은 나중",
        "operationId": "create_analysis_api_v0_analyses_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "submissionId:attemptNo. 중복 요청 판별",
              "title": "Idempotency-Key"
            },
            "description": "submissionId:attemptNo. 중복 요청 판별"
          },
          {
            "name": "x-trace-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "분산 추적 ID",
              "title": "X-Trace-Id"
            },
            "description": "분산 추적 ID"
          },
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisAccepted"
                }
              }
            }
          },
          "422": {
            "description": "요청 스키마 위반",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "description": "POST /api/v0/analyses 요청 본문 ",
                "properties": {
                  "attemptId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "Spring 측 측정수행 키(에코용)",
                    "title": "Attemptid"
                  },
                  "submissionId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Submissionid"
                  },
                  "callbackUrl": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "완료 통지 수신 주소. 현재는 수용만 하고 전송은 미구현",
                    "title": "Callbackurl"
                  },
                  "method": {
                    "enum": [
                      "GITHUB_URL",
                      "ZIP_WITH_GITLOG"
                    ],
                    "title": "Method",
                    "type": "string"
                  },
                  "source": {
                    "properties": {
                      "repoUrl": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ],
                        "description": "method=GITHUB_URL일 때 필수. 공개 레포만 지원",
                        "title": "Repourl"
                      },
                      "branch": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ],
                        "description": "생략 시 기본 브랜치",
                        "title": "Branch"
                      }
                    },
                    "title": "AnalysisSource",
                    "type": "object"
                  },
                  "extractionScope": {
                    "default": "TOTAL",
                    "enum": [
                      "TOTAL",
                      "OWN_COMMIT"
                    ],
                    "title": "Extractionscope",
                    "type": "string"
                  },
                  "commitEmail": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "OWN_COMMIT일 때 필수",
                    "title": "Commitemail"
                  },
                  "questionBudget": {
                    "default": 3,
                    "description": "계획 문제 수",
                    "minimum": 1,
                    "title": "Questionbudget",
                    "type": "integer"
                  },
                  "focusItems": {
                    "description": "강사 지정 초점 후보. 비면 AI 자율 선정",
                    "items": {
                      "description": "강사가 지정한 질문 초점 후보. Spring question_focus_item 행 대응.\n\nAI는 이 중 하나의 id를 골라 problem.questionFocusItemId로 돌려준다.\nPK가 랜덤 UUID라 AI 코드에 값을 박을 수 없어서 후보를 받아 고르는 방식이다(C-1).",
                      "properties": {
                        "id": {
                          "title": "Id",
                          "type": "string"
                        },
                        "name": {
                          "title": "Name",
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "name"
                      ],
                      "title": "FocusItem",
                      "type": "object"
                    },
                    "title": "Focusitems",
                    "type": "array"
                  },
                  "requirements": {
                    "description": "[{requirementId, text}] P/F 판정 대상",
                    "items": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "title": "Requirements",
                    "type": "array"
                  },
                  "teaches": {
                    "description": "[{id, label, unitId, sourcePages}] 교안 참조용",
                    "items": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "title": "Teaches",
                    "type": "array"
                  },
                  "curriculumId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Curriculumid"
                  },
                  "modelCode": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "description": "생략 시 서버 기본값. operator가 고른다",
                    "title": "Modelcode"
                  }
                },
                "required": [
                  "method"
                ],
                "title": "AnalysisRequest",
                "type": "object"
              },
              "example": {
                "method": "GITHUB_URL",
                "source": {
                  "repoUrl": "https://github.com/owner/repo",
                  "branch": "main"
                },
                "extractionScope": "TOTAL",
                "questionBudget": 3
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "payload",
                  "file"
                ],
                "properties": {
                  "payload": {
                    "type": "string",
                    "contentMediaType": "application/json",
                    "contentSchema": {
                      "description": "POST /api/v0/analyses 요청 본문 ",
                      "properties": {
                        "attemptId": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "description": "Spring 측 측정수행 키(에코용)",
                          "title": "Attemptid"
                        },
                        "submissionId": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "title": "Submissionid"
                        },
                        "callbackUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "description": "완료 통지 수신 주소. 현재는 수용만 하고 전송은 미구현",
                          "title": "Callbackurl"
                        },
                        "method": {
                          "enum": [
                            "GITHUB_URL",
                            "ZIP_WITH_GITLOG"
                          ],
                          "title": "Method",
                          "type": "string"
                        },
                        "source": {
                          "properties": {
                            "repoUrl": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ],
                              "description": "method=GITHUB_URL일 때 필수. 공개 레포만 지원",
                              "title": "Repourl"
                            },
                            "branch": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ],
                              "description": "생략 시 기본 브랜치",
                              "title": "Branch"
                            }
                          },
                          "title": "AnalysisSource",
                          "type": "object"
                        },
                        "extractionScope": {
                          "default": "TOTAL",
                          "enum": [
                            "TOTAL",
                            "OWN_COMMIT"
                          ],
                          "title": "Extractionscope",
                          "type": "string"
                        },
                        "commitEmail": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "description": "OWN_COMMIT일 때 필수",
                          "title": "Commitemail"
                        },
                        "questionBudget": {
                          "default": 3,
                          "description": "계획 문제 수",
                          "minimum": 1,
                          "title": "Questionbudget",
                          "type": "integer"
                        },
                        "focusItems": {
                          "description": "강사 지정 초점 후보. 비면 AI 자율 선정",
                          "items": {
                            "description": "강사가 지정한 질문 초점 후보. Spring question_focus_item 행 대응.\n\nAI는 이 중 하나의 id를 골라 problem.questionFocusItemId로 돌려준다.\nPK가 랜덤 UUID라 AI 코드에 값을 박을 수 없어서 후보를 받아 고르는 방식이다(C-1).",
                            "properties": {
                              "id": {
                                "title": "Id",
                                "type": "string"
                              },
                              "name": {
                                "title": "Name",
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name"
                            ],
                            "title": "FocusItem",
                            "type": "object"
                          },
                          "title": "Focusitems",
                          "type": "array"
                        },
                        "requirements": {
                          "description": "[{requirementId, text}] P/F 판정 대상",
                          "items": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "title": "Requirements",
                          "type": "array"
                        },
                        "teaches": {
                          "description": "[{id, label, unitId, sourcePages}] 교안 참조용",
                          "items": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "title": "Teaches",
                          "type": "array"
                        },
                        "curriculumId": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "title": "Curriculumid"
                        },
                        "modelCode": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "description": "생략 시 서버 기본값. operator가 고른다",
                          "title": "Modelcode"
                        }
                      },
                      "required": [
                        "method"
                      ],
                      "title": "AnalysisRequest",
                      "type": "object"
                    },
                    "description": "요청 JSON을 문자열로. 구조는 AnalysisRequest",
                    "example": "{\"method\":\"ZIP_WITH_GITLOG\",\"extractionScope\":\"TOTAL\"}"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "제출 ZIP"
                  }
                }
              },
              "encoding": {
                "payload": {
                  "contentType": "application/json"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/analyses/{job_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "분석 상태·결과 조회 (P02)",
        "description": "분석 job의 현재 상태와 결과를 돌려준다.\n\n경로의 {job_id}가 함수 파라미터 job_id로 자동 연결된다. 이름이 같아야 한다.\n\njob 저장소가 인메모리라 프로세스가 재시작되면 404가 난다.\n그때는 Spring이 분석을 다시 요청하면 된다(FastAPI는 상태의 소유자가 아니다).",
        "operationId": "get_analysis_api_v0_analyses__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "x-internal-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJobStatus"
                }
              }
            }
          },
          "404": {
            "description": "모르는 job_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiUsage": {
        "properties": {
          "featureCode": {
            "type": "string",
            "enum": [
              "CODE_ANALYSIS",
              "QUESTION_GENERATION",
              "GRADING",
              "CURRICULUM_ANALYSIS",
              "SUMMARY_DRAFT"
            ],
            "title": "Featurecode"
          },
          "modelCode": {
            "type": "string",
            "title": "Modelcode",
            "description": "Spring이 ai_model에서 model_id를 조회한다"
          },
          "sourceType": {
            "type": "string",
            "title": "Sourcetype",
            "description": "값 목록은 백엔드 확정 대기(C-4)"
          },
          "sourceId": {
            "type": "string",
            "title": "Sourceid",
            "description": "작업 PK. 분석이면 analysisId"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          },
          "traceId": {
            "type": "string",
            "title": "Traceid",
            "description": "요청 헤더 X-Trace-Id를 그대로 잇는다"
          },
          "idempotencyKey": {
            "type": "string",
            "title": "Idempotencykey",
            "description": "{sourceId}:{sourceType}:{attemptNo} 형식"
          },
          "inputTokenCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Inputtokencount"
          },
          "outputTokenCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Outputtokencount"
          },
          "cachedTokenCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cachedtokencount",
            "default": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "SUCCEEDED",
              "FAILED",
              "PARTIAL"
            ],
            "title": "Status"
          },
          "failureCode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "TIMEOUT",
                  "RATE_LIMITED",
                  "PROVIDER_ERROR",
                  "INVALID_JSON",
                  "CONTEXT_OVERFLOW"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Failurecode"
          },
          "latencyMs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Latencyms"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time",
            "title": "Occurredat"
          }
        },
        "type": "object",
        "required": [
          "featureCode",
          "modelCode",
          "sourceType",
          "sourceId",
          "requestId",
          "traceId",
          "idempotencyKey",
          "inputTokenCount",
          "outputTokenCount",
          "status",
          "latencyMs",
          "occurredAt"
        ],
        "title": "AiUsage",
        "description": "LLM 호출 한 번의 기록. DB ai_usage 한 행에 대응한다."
      },
      "AnalysisAccepted": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "status": {
            "type": "string",
            "const": "QUEUED",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "status"
        ],
        "title": "AnalysisAccepted",
        "description": "202 응답. 점수만 알리고 결과는 폴링으로 가져감 "
      },
...
      "AnalysisJobStatus": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "attemptId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attemptid"
          },
          "submissionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submissionid"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "PARTIAL",
              "FAILED"
            ],
            "title": "Status"
          },
          "failureReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failurereason",
            "description": "FAILED일 때만 채워진다"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "SUCCEEDED·PARTIAL일 때만"
          },
          "aiUsage": {
            "items": {
              "$ref": "#/components/schemas/AiUsage"
            },
            "type": "array",
            "title": "Aiusage"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "status"
        ],
        "title": "AnalysisJobStatus",
        "description": "GET /analyses/{jobId} 응답.\n\nstatus 값은 DB analysis_job.status의 CHECK 제약과 같다(PLAN §3).\nANALYZING·READY는 다른 테이블의 값이라 여기 쓰면 Spring INSERT가 깨진다."
      },
      "AnalysisResult": {
        "properties": {
          "snapshotId": {
            "type": "string",
            "title": "Snapshotid",
            "description": "Spring code_snapshot 행의 키"
          },
          "snapshotMeta": {
            "$ref": "#/components/schemas/SnapshotMeta"
          },
          "appliedScope": {
            "type": "string",
            "enum": [
              "TOTAL",
              "OWN_COMMIT"
            ],
            "title": "Appliedscope"
          },
          "scopeFallback": {
            "type": "boolean",
            "title": "Scopefallback",
            "description": "요청 범위를 못 지켜 TOTAL로 물러났는지"
          },
          "fallbackReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallbackreason"
          },
          "commitSha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commitsha"
          },
          "analysisDocument": {
            "$ref": "#/components/schemas/AnalysisDocument",
            "description": "code_analysis.analysis_document 대응. Markdown이 아니라 구조화 JSON이다"
          },
          "requirementResults": {
            "items": {
              "$ref": "#/components/schemas/RequirementResult"
            },
            "type": "array",
            "title": "Requirementresults"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/Problem"
            },
            "type": "array",
            "title": "Problems"
          },
          "questionCountPlanned": {
            "type": "integer",
            "title": "Questioncountplanned",
            "description": "계획된 질문 수. 유효 문제가 적으면 축소된다"
          }
        },
        "type": "object",
        "required": [
          "snapshotId",
          "snapshotMeta",
          "appliedScope",
          "scopeFallback",
          "analysisDocument",
          "questionCountPlanned"
        ],
        "title": "AnalysisResult",
        "description": "분석이 성공했을 때의 결과 본문.\n\n문제·질문·힌트가 전부 여기 실린다. 질문과 힌트는 분석 때 만들어 동결하므로\n세션 시작에는 AI 호출이 없다 — Spring이 이 응답을 저장해두고 꺼내 쓴다."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SnapshotMeta": {
        "properties": {
          "contentHash": {
            "type": "string",
            "title": "Contenthash",
            "description": "sha256 hex 64자"
          },
          "fileCount": {
            "type": "integer",
            "title": "Filecount"
          },
          "byteCount": {
            "type": "integer",
            "title": "Bytecount"
          }
        },
        "type": "object",
        "required": [
          "contentHash",
          "fileCount",
          "byteCount"
        ],
        "title": "SnapshotMeta",
        "description": "code_snapshot 테이블 대응. 코드 원문 저장하지 않고 메타만 주기 "
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}